Needle Wobble correction using CNC domain
There are several changes to the needle wobble algorithms.
The Offset2Method_button_Click function was the first to be changed
and has an extensive discussion at the top.  The biggest change is to
store the current CNC position in the needle CalibrationPoints array.
The CNC position of the down camera over the up camera is calculated
(we can't actually move there) and is used to calculate the Needle Offset
for every calibration point.  The interpolation routines still work.
The other big change is to programmatically move the needle to the
center of the camera for every measurement.  This should give the highest
accuracy in the CNC domain.  Finally, the "Set" button has been removed
for the Up Camera location.  It is set as part of the Needle Setup.