Workaround for Issue #56: Pump is not switched on when placing parts!
Recently I updated my old fork to the newest version. Now I have the problem that when I start placing parts the pump is not switched on. After start of program Pump On & Vacuum On(solenoid valve) is marked as on, so the program thinks that the pump is already switched on. To get it work I need to manually switch off the pump & solenoid valve in the Basic Setup tab once.

Since I am not shure about the root cause I simply removed the following settings:
 VacuumIsOn = true
PumpIsOn = true