Workarounf for Issue #53: Pressing stop while nozzle is placing an object puts the machine in a unrecoverable state.
The probleme here is that if the Stop button is pressed and the actual action was aborted, the status variable AbortPlacement is still true. Because of this, starting a new action which includes moving the machine will be always aborted (return false).

I am not shure if this covers all the problems from issue #53, but for me it solved the problem.